Done: MailComposerModel
+ conformance to MFMailComposeViewControllerDelegate
